sharding by other column

+++ sharding-jdbc-example/spring-namespace-example/spring-namespace-mybatis-example/src/main/java/io/shardingsphere/example/spring/namespace/mybatis/fixtrue/algorithm/PreciseModuloTableShardingAlgorithm.java
@@ -22,10 +22,10 @@ import io.shardingsphere.core.api.algorithm.sharding.standard.PreciseShardingAlg
 
 import java.util.Collection;
 
-public final class PreciseModuloTableShardingAlgorithm implements PreciseShardingAlgorithm<Long> {
+public final class PreciseModuloTableShardingAlgorithm implements PreciseShardingAlgorithm<Integer> {
     
     @Override
-    public String doSharding(final Collection<String> availableTargetNames, final PreciseShardingValue<Long> shardingValue) {
+    public String doSharding(final Collection<String> availableTargetNames, final PreciseShardingValue<Integer> shardingValue) {
         for (String each : availableTargetNames) {
             if (each.endsWith(shardingValue.getValue() % 2 + "")) {
                 return each;
